requirements.txt
asgiref==3.5.2
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.6
sqlparse==0.4.3
python-dotenv==0.21.0
drf-yasg==1.21.4
Pillow==9.2.0
django-debug-toolbar==3.8.1
django-phonenumbers==1.0.1 
django-phonenumber-field==7.0.0 